formate

+++ im2latex/model/components/beam_search_decoder_cell.py
@@ -43,7 +43,7 @@ class BeamSearchDecoderOutput(collections.namedtuple(
 class BeamSearchDecoderCell(object):
 
     def __init__(self, embeddings, cell, batch_size, start_token, end_token,
-            beam_size=5, div_gamma=1, div_prob=0):
+                 beam_size=5, div_gamma=1, div_prob=0):
         """Initializes parameters for Beam Search
 
         Args:
@@ -66,67 +66,59 @@ class BeamSearchDecoderCell(object):
         self._dim_embeddings = embeddings.shape[-1].value
         self._batch_size = batch_size
         self._start_token = start_token
-        self._beam_size  = beam_size
+        self._beam_size = beam_size
         self._end_token = end_token
         self._vocab_size = embeddings.shape[0].value
         self._div_gamma = float(div_gamma)
         self._div_prob = float(div_prob)
 
-
     @property
     def output_dtype(self):
         """Needed for custom dynamic_decode for the TensorArray of results"""
         return BeamSearchDecoderOutput(logits=self._cell.output_dtype,
-                ids=tf.int32, parents=tf.int32)
-
+                                       ids=tf.int32, parents=tf.int32)
 
     @property
     def final_output_dtype(self):
         """For the finalize method"""
         return DecoderOutput(logits=self._cell.output_dtype, ids=tf.int32)
 
-
     @property
     def state_size(self):
         return BeamSearchDecoderOutput(
-                logits=tf.TensorShape([self._beam_size, self._vocab_size]),
-                ids=tf.TensorShape([self._beam_size]),
-                parents=tf.TensorShape([self._beam_size]))
-
+            logits=tf.TensorShape([self._beam_size, self._vocab_size]),
+            ids=tf.TensorShape([self._beam_size]),
+            parents=tf.TensorShape([self._beam_size]))
 
     @property
     def final_output_size(self):
         return DecoderOutput(logits=tf.TensorShape([self._beam_size,
-                self._vocab_size]), ids=tf.TensorShape([self._beam_size]))
-
+                                                    self._vocab_size]), ids=tf.TensorShape([self._beam_size]))
 
     def initial_state(self):
         """Returns initial state for the decoder"""
         # cell initial state
         cell_state = self._cell.initial_state()
         cell_state = nest.map_structure(lambda t: tile_beam(t,
-                self._beam_size), cell_state)
+                                                            self._beam_size), cell_state)
 
         # prepare other initial states
-        log_probs =  tf.zeros([self._batch_size, self._beam_size],
-                dtype=self._cell.output_dtype)
+        log_probs = tf.zeros([self._batch_size, self._beam_size],
+                             dtype=self._cell.output_dtype)
 
         return BeamSearchDecoderCellState(cell_state, log_probs)
 
-
     def initial_inputs(self):
         return tf.tile(tf.reshape(self._start_token,
-                [1, 1, self._dim_embeddings]),
-                multiples=[self._batch_size, self._beam_size, 1])
-
+                                  [1, 1, self._dim_embeddings]),
+                       multiples=[self._batch_size, self._beam_size, 1])
 
     def initialize(self):
         initial_state = self.initial_state()
         initial_inputs = self.initial_inputs()
         initial_finished = tf.zeros(shape=[self._batch_size, self._beam_size],
-                dtype=tf.bool)
+                                    dtype=tf.bool)
         return initial_state, initial_inputs, initial_finished
-
 
     def step(self, time, state, embedding, finished):
         """
@@ -147,7 +139,7 @@ class BeamSearchDecoderCell(object):
         # split batch and beam dimension before beam search logic
         new_logits = split_batch_beam(logits, self._beam_size)
         new_cell_state = nest.map_structure(
-                lambda t: split_batch_beam(t, self._beam_size), new_cell_state)
+            lambda t: split_batch_beam(t, self._beam_size), new_cell_state)
 
         # compute log probs of the step
         # shape = [batch_size, beam_size, vocab_size]
@@ -157,15 +149,15 @@ class BeamSearchDecoderCell(object):
         # shape = [batch_size, beam_size, vocab_size]
         log_probs = tf.expand_dims(state.log_probs, axis=-1) + step_log_probs
         log_probs = add_div_penalty(log_probs, self._div_gamma, self._div_prob,
-                self._batch_size, self._beam_size, self._vocab_size)
+                                    self._batch_size, self._beam_size, self._vocab_size)
 
         # compute the best beams
         # shape =  (batch_size, beam_size * vocab_size)
         log_probs_flat = tf.reshape(log_probs,
-                [self._batch_size, self._beam_size * self._vocab_size])
+                                    [self._batch_size, self._beam_size * self._vocab_size])
         # if time = 0, consider only one beam, otherwise beams are equal
         log_probs_flat = tf.cond(time > 0, lambda: log_probs_flat,
-                lambda: log_probs[:, 0])
+                                 lambda: log_probs[:, 0])
         new_probs, indices = tf.nn.top_k(log_probs_flat, self._beam_size)
 
         # of shape [batch_size, beam_size]
@@ -177,24 +169,22 @@ class BeamSearchDecoderCell(object):
 
         # compute end of beam
         finished = gather_helper(finished, new_parents,
-                self._batch_size, self._beam_size)
+                                 self._batch_size, self._beam_size)
         new_finished = tf.logical_or(finished,
-                tf.equal(new_ids, self._end_token))
+                                     tf.equal(new_ids, self._end_token))
 
         new_cell_state = nest.map_structure(
-                lambda t: gather_helper(t, new_parents, self._batch_size,
-                self._beam_size), new_cell_state)
-
+            lambda t: gather_helper(t, new_parents, self._batch_size,
+                                    self._beam_size), new_cell_state)
 
         # create new state of decoder
-        new_state  = BeamSearchDecoderCellState(cell_state=new_cell_state,
-                log_probs=new_probs)
+        new_state = BeamSearchDecoderCellState(cell_state=new_cell_state,
+                                               log_probs=new_probs)
 
         new_output = BeamSearchDecoderOutput(logits=new_logits, ids=new_ids,
-                parents=new_parents)
+                                             parents=new_parents)
 
         return (new_output, new_state, new_embedding, new_finished)
-
 
     def finalize(self, final_outputs, final_state):
         """
@@ -210,7 +200,7 @@ class BeamSearchDecoderCell(object):
         # reverse the time dimension
         maximum_iterations = tf.shape(final_outputs.ids)[0]
         final_outputs = nest.map_structure(lambda t: tf.reverse(t, axis=[0]),
-                final_outputs)
+                                           final_outputs)
 
         # initial states
         def create_ta(d):
@@ -218,10 +208,10 @@ class BeamSearchDecoderCell(object):
 
         initial_time = tf.constant(0, dtype=tf.int32)
         initial_outputs_ta = nest.map_structure(create_ta,
-                self.final_output_dtype)
+                                                self.final_output_dtype)
         initial_parents = tf.tile(
-                tf.expand_dims(tf.range(self._beam_size), axis=0),
-                multiples=[self._batch_size, 1])
+            tf.expand_dims(tf.range(self._beam_size), axis=0),
+            multiples=[self._batch_size, 1])
 
         def condition(time, outputs_ta, parents):
             return tf.less(time, maximum_iterations)
@@ -233,31 +223,29 @@ class BeamSearchDecoderCell(object):
 
             # extract the entries corresponding to parents
             new_state = nest.map_structure(
-                    lambda t: gather_helper(t, parents, self._batch_size,
-                    self._beam_size), input_t)
+                lambda t: gather_helper(t, parents, self._batch_size,
+                                        self._beam_size), input_t)
 
             # create new output
             new_output = DecoderOutput(logits=new_state.logits,
-                    ids=new_state.ids)
+                                       ids=new_state.ids)
 
             # write beam ids
             outputs_ta = nest.map_structure(lambda ta, out: ta.write(time, out),
-                    outputs_ta, new_output)
+                                            outputs_ta, new_output)
 
             return (time + 1), outputs_ta, parents
 
-        res = tf.while_loop(
-                condition,
-                body,
-                loop_vars=[initial_time, initial_outputs_ta, initial_parents],
-                back_prop=False)
+        res = tf.while_loop(condition, body,
+                            loop_vars=[initial_time, initial_outputs_ta, initial_parents],
+                            back_prop=False)
 
         # unfold and stack the structure from the nested tas
         final_outputs = nest.map_structure(lambda ta: ta.stack(), res[1])
 
         # reverse time step
         final_outputs = nest.map_structure(lambda t: tf.reverse(t, axis=[0]),
-                final_outputs)
+                                           final_outputs)
 
         return DecoderOutput(logits=final_outputs.logits, ids=final_outputs.ids)
 
@@ -268,7 +256,7 @@ def sample_bernoulli(p, s):
 
 
 def add_div_penalty(log_probs, div_gamma, div_prob, batch_size, beam_size,
-        vocab_size):
+                    vocab_size):
     """Adds penalty to beam hypothesis following this paper by Li et al. 2016
     "A Simple, Fast Diverse Decoding Algorithm for Neural Generation"
 
@@ -279,22 +267,21 @@ def add_div_penalty(log_probs, div_gamma, div_prob, batch_size, beam_size,
         div_prob: (float) adds penalty with proba div_prob
 
     """
-    if div_gamma is None or div_prob is None: return log_probs
-    if div_gamma == 1. or div_prob == 0.: return log_probs
+    if div_gamma is None or div_prob is None:
+        return log_probs
+    if div_gamma == 1. or div_prob == 0.:
+        return log_probs
 
     # 1. get indices that would sort the array
     top_probs, top_inds = tf.nn.top_k(log_probs, k=vocab_size, sorted=True)
     # 2. inverse permutation to get rank of each entry
     top_inds = tf.reshape(top_inds, [-1, vocab_size])
     index_rank = tf.map_fn(tf.invert_permutation, top_inds, back_prop=False)
-    index_rank = tf.reshape(index_rank, shape=[batch_size, beam_size,
-            vocab_size])
+    index_rank = tf.reshape(index_rank, shape=[batch_size, beam_size, vocab_size])
     # 3. compute penalty
     penalties = tf.log(div_gamma) * tf.cast(index_rank, log_probs.dtype)
     # 4. only apply penalty with some probability
-    apply_penalty = tf.cast(
-            sample_bernoulli(div_prob, [batch_size, beam_size, vocab_size]),
-            penalties.dtype)
+    apply_penalty = tf.cast(sample_bernoulli(div_prob, [batch_size, beam_size, vocab_size]), penalties.dtype)
     penalties *= apply_penalty
 
     return log_probs + penalties
@@ -318,8 +305,7 @@ def merge_batch_beam(t):
     elif t.shape.ndims == 3:
         return tf.reshape(t, [batch_size*beam_size, t.shape[-1].value])
     elif t.shape.ndims == 4:
-        return tf.reshape(t, [batch_size*beam_size, t.shape[-2].value,
-                t.shape[-1].value])
+        return tf.reshape(t, [batch_size*beam_size, t.shape[-2].value, t.shape[-1].value])
     else:
         raise NotImplementedError
 
@@ -338,8 +324,7 @@ def split_batch_beam(t, beam_size):
     elif t.shape.ndims == 2:
         return tf.reshape(t, [-1, beam_size, t.shape[-1].value])
     elif t.shape.ndims == 3:
-        return tf.reshape(t, [-1, beam_size, t.shape[-2].value,
-                t.shape[-1].value])
+        return tf.reshape(t, [-1, beam_size, t.shape[-2].value, t.shape[-1].value])
     else:
         raise NotImplementedError
 
@@ -375,7 +360,7 @@ def mask_probs(probs, end_token, finished):
     # one hot of shape [vocab_size]
     vocab_size = probs.shape[-1].value
     one_hot = tf.one_hot(end_token, vocab_size, on_value=0.,
-            off_value=probs.dtype.min, dtype=probs.dtype)
+                         off_value=probs.dtype.min, dtype=probs.dtype)
     # expand dims of shape [batch_size, beam_size, 1]
     finished = tf.expand_dims(tf.cast(finished, probs.dtype), axis=-1)
 
@@ -392,9 +377,9 @@ def gather_helper(t, indices, batch_size, beam_size):
         new_t: tensor w shape as t but new_t[:, i] = t[:, new_parents[:, i]]
 
     """
-    range_  = tf.expand_dims(tf.range(batch_size) * beam_size, axis=1)
+    range_ = tf.expand_dims(tf.range(batch_size) * beam_size, axis=1)
     indices = tf.reshape(indices + range_, [-1])
-    output  = tf.gather(
+    output = tf.gather(
         tf.reshape(t, [batch_size*beam_size, -1]),
         indices)
